+++ src/locales/en_US.ts
@@ -1141,7 +1141,7 @@ const strings = {
   exchange_slow: 'Locating a swap is taking longer than usual.\nPlease wait...',
   wallet_list_modal_header_parent: 'Parent Wallet',
   wallet_list_modal_header_mru: 'Most Recent Wallets',
-  wallet_list_modal_header_all: 'All Wallets',
+  wallet_list_modal_header_other: 'Other Wallets',
   wallet_list_modal_creating_wallet: 'Creating Wallet. Please Wait',
   wallet_list_modal_enabling_token: 'Enabling Token. Please Wait',
   wallet_list_modal_confirm_s_bank_withdrawal: